Add user ID check on oldState
PR #75 introduced a bug that caused the bot to leave whenever a user left the VC, regardless of if the bot was alone or not. The admin kick check correctly looked for the sudden loss of connection, but did not check if the voiceStateUpdate was in the context of the bot user, so it would read every user's update. It now checks the ID of the disconnected member to see if it's the bot that left. 
Fixes #77